separated bookdetails on comments table and card details
