+++ WBooks/Screens/BookDetails/CardDetails/CardDetailsViewModel.swift
@@ -1,30 +1,24 @@
 //
-//  BookDetailsViewModel.swift
+//  CardDetailsViewModel.swift
 //  WBooks
 //
-//  Created by Nacho 2 on 05/04/2020.
+//  Created by Nacho 2 on 10/05/2020.
 //  Copyright © 2020 Wolox. All rights reserved.
 //
 
 import Foundation
 
-class BookDetailsViewModel {
+class CardDetailsViewModel {
     private let _bookRepository = BookRepository()
     private let _userRepository = UserRepository()
     private var _bookModel: Book
-    private var commentsList: [Comment]
     
     init(bookModel: Book) {
         _bookModel = bookModel
-        commentsList = []
     }
     
-    var numberOfComments: Int {
-        return commentsList.count
-    }
-    
-    var isEmptyComments: Bool {
-        return commentsList.isEmpty
+    var id: Int {
+        _bookModel.id
     }
     
     var status: BookStatus {
@@ -51,13 +45,9 @@ class BookDetailsViewModel {
         guard let imageName = _bookModel.image, imageName.isNotEmpty else { return nil }
         return  imageName
     }
-    
-    func getCommentCellViewModel(at indexPath: IndexPath) -> CommentCellViewModel {
-        CommentCellViewModel(commentModel: commentsList[indexPath.row])
-    }
 }
 
-extension BookDetailsViewModel {
+extension CardDetailsViewModel {
     func rentBook(onSuccess: @escaping () -> Void, onError: @escaping () -> Void) {
         let rentSuccess = {
             self._bookModel.status = BookStatus.unavailable.rawValue
@@ -72,16 +62,9 @@ extension BookDetailsViewModel {
         let rentModel = Rent(bookId: _bookModel.id, from: from, to: to)
         _userRepository.rentBook(rentModel: rentModel, onSuccess: rentSuccess, onError: rentError)
     }
-    
-    func getBookComments(onSuccess: @escaping () -> Void, onError: @escaping () -> Void) {
-        let commentsSuccess: ([Comment]) -> Void = { comments in
-            self.commentsList = comments
-            onSuccess()
-        }
-        let commentsError: (Error) -> Void = { error in
-            onError()
-            print(error)
-        }
-        _bookRepository.fetchComments(bookId: self._bookModel.id, onSuccess: commentsSuccess, onError: commentsError)
-    }
+}
+
+enum BookStatus: String {
+    case available = "Available"
+    case unavailable = "Unavailable"
 }
