+++ WBooks/Screens/BookDetails/CommentsTable/CommentsTableView.xib
@@ -1,0 +1,41 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="15505" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina6_1" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="15510"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <view contentMode="scaleAspectFit" id="iN0-l3-epB" userLabel="Comments Table View" customClass="CommentsTableView" customModule="WBooks" customModuleProvider="target">
+            <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
+            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+            <subviews>
+                <tableView clipsSubviews="YES" contentMode="scaleAspectFit" alwaysBounceVertical="YES" style="plain" separatorStyle="none" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="28" sectionFooterHeight="28" translatesAutoresizingMaskIntoConstraints="NO" id="thq-5z-ttQ" userLabel="Comments Table">
+                    <rect key="frame" x="20" y="398" width="374" height="464"/>
+                    <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
+                </tableView>
+                <activityIndicatorView opaque="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="750" verticalHuggingPriority="750" style="medium" id="x37-TO-yu4" userLabel="CommentsLoading">
+                    <rect key="frame" x="20" y="608" width="374" height="44"/>
+                    <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+                </activityIndicatorView>
+            </subviews>
+            <color key="backgroundColor" red="1" green="1" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+            <constraints>
+                <constraint firstItem="thq-5z-ttQ" firstAttribute="trailing" secondItem="x37-TO-yu4" secondAttribute="trailing" id="1S0-DE-e2K"/>
+                <constraint firstItem="thq-5z-ttQ" firstAttribute="leading" secondItem="x37-TO-yu4" secondAttribute="leading" id="DQc-uV-eJd"/>
+                <constraint firstItem="thq-5z-ttQ" firstAttribute="centerY" secondItem="x37-TO-yu4" secondAttribute="centerY" id="EcP-Bm-stJ"/>
+                <constraint firstItem="thq-5z-ttQ" firstAttribute="bottom" secondItem="vUN-kp-3ea" secondAttribute="bottom" id="kpp-ji-Ued"/>
+            </constraints>
+            <viewLayoutGuide key="safeArea" id="vUN-kp-3ea"/>
+            <connections>
+                <outlet property="commentsLoadingIndicator" destination="x37-TO-yu4" id="5Zg-1t-Aux"/>
+                <outlet property="commentsTable" destination="thq-5z-ttQ" id="Ug7-UA-pwf"/>
+            </connections>
+            <point key="canvasLocation" x="34.782608695652179" y="-207.58928571428569"/>
+        </view>
+    </objects>
+</document>
