+++ WBooks/Screens/Library/LibraryViewController.swift
@@ -96,7 +96,7 @@ extension LibraryViewController: UITableViewDataSource {
     }
     
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        let bookDetailsVM = BookDetailsViewModel(bookModel: _libraryViewModel.booksList[indexPath.row])
+        let bookDetailsVM = CardDetailsViewModel(bookModel: _libraryViewModel.booksList[indexPath.row])
         let bookVC = BookDetailsViewController(viewModel: bookDetailsVM)
         navigationController?.pushViewController(bookVC, animated: true)
     }
